+++ kernel_tuner/runners/sequential_brute_force.py
@@ -9,7 +9,7 @@ from kernel_tuner.core import *
 def run(kernel_name, original_kernel, problem_size, arguments,
         tune_params, parameter_space, grid_div_x, grid_div_y,
         answer, atol, verbose,
-        lang, device, platform, cmem_args):
+        lang, device, platform, cmem_args, compiler_options=None):
     """ Iterate through the entire parameter space using a single Python process
 
     :param kernel_name: The name of the kernel in the code.
@@ -69,7 +69,7 @@ def run(kernel_name, original_kernel, problem_size, arguments,
 
     #detect language and create device function interface
     lang = detect_language(lang, original_kernel)
-    dev = get_device_interface(lang, device, platform)
+    dev = get_device_interface(lang, device, platform, compiler_options)
 
     #move data to the GPU
     gpu_args = dev.ready_argument_list(arguments)